backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
torch
torchvision
pillow
git+https://github.com/openai/CLIP.git
